==> Nootbook/01_defining_the_problem.ipynb <==
# Holiday Package Prediction

### Problem Statement:
"Travel.Com" company wants to enable and establish a viable business model to expand the customer base. One of the ways to expand the customer base is to introduce a new offering of packages. Currently, there are 5 types of packages the company is offering - Basic, Standard, Deluxe, Super Deluxe, King. Looking at the data of the last year, we observed that 18% of the customers purchased the packages. However, the marketing cost was quite high because customers were contacted at random without looking at the available information. The company is now planning to launch a new product i.e. Wellness Tourism Package. Wellness Tourism is defined as Travel that allows the traveler to maintain, enhance or kick-start a healthy lifestyle, and support or increase one's sense of well-being.However, this time company wants to harness the available data of existing and potential customers to make the marketing expenditure more efficient.

We need to analyze the customers' data and information to provide recommendations to the Policy Maker and Marketing Team and also build a model to predict the potential customer who is going to purchase the newly introduced travel package.

### Data Collection
Data is Collected form https://www.kaggle.com/code/nisharbadi/holiday-package-prediction/input, Data contains 4888 row and 20 columns

import pandas as pd 
df = pd.read_csv(r"D:\Amaan_Behlim\01_Udemy_Course\05_Machine_Learning\09_ADABoost\Data\Raw\Travel.csv")
df
