==> Nootbook/02_data_cleaning.ipynb <==
# Data Cleaning
1. Handling Missing Values
2. Handling Duplicates
3. Check Datatypee
4. Understand the dataset 

import pandas as pd 
import numpy as np
import matplotlib as plt
import seaborn as sns

df = pd.read_csv(r"D:\Amaan_Behlim\01_Udemy_Course\05_Machine_Learning\09_ADABoost\Data\Raw\Travel.csv")
df.head()

df.info()

df.describe()

1. Handling Missing Values

pd.DataFrame(df.isnull().sum())

df.head()

#Age
df['Age'] = df['Age'].fillna(df['Age'].mean())

#TypeofContact 
df['TypeofContact'] = df['TypeofContact'].fillna(df['TypeofContact'].mode()[0])

#DurationOfPitch
df['DurationOfPitch'] = df['DurationOfPitch'].fillna(df['DurationOfPitch'].median())

#NumberOfFollowups
df['NumberOfFollowups'] = df['NumberOfFollowups'].fillna(df['NumberOfFollowups'].mode()[0])

#PreferredPropertyStar
df['PreferredPropertyStar'] = df['PreferredPropertyStar'].fillna(df['PreferredPropertyStar'].mode()[0])

#NumberOfTrips
df['NumberOfTrips'] = df['NumberOfTrips'].fillna(df['NumberOfTrips'].median())

#NumberOfChildrenVisiting
df['NumberOfChildrenVisiting'] = df['NumberOfChildrenVisiting'].fillna(df['NumberOfChildrenVisiting'].mode()[0])

#MonthlyIncome
df['MonthlyIncome'] = df['MonthlyIncome'].fillna(df['MonthlyIncome'].mean())

df.isnull().sum()

2. Handling Duplicates 

df.duplicated().sum()

3. Checking datatype

print(df['TypeofContact'].value_counts())
print('\n')
print(df['Occupation'].value_counts())
print('\n')
print(df['Gender'].value_counts())

print('\n')
print(df['ProductPitched'].value_counts())
print('\n')
print(df['MaritalStatus'].value_counts())
print('\n')
print(df['Designation'].value_counts())

df['MaritalStatus'] = df['MaritalStatus'].replace('Single', 'Unmarried')
df['Designation'] = df['Designation'].replace('AVP', 'Assistant Vice President') 
df['Designation'] = df['Designation'].replace('VP', 'Vice President') 
df['Gender'] = df['Gender'].replace('Fe Male', 'Female') 
df['Occupation'] = df['Occupation'].replace('Small Business', 'Business')
df['Occupation'] = df['Occupation'].replace('Large Business', 'Business')


print(df['TypeofContact'].value_counts())
print('\n')
print(df['Occupation'].value_counts())
print('\n')
print(df['Gender'].value_counts())

print('\n')
print(df['ProductPitched'].value_counts())
print('\n')
print(df['MaritalStatus'].value_counts())
print('\n')
print(df['Designation'].value_counts())

df.to_csv(r"D:\Amaan_Behlim\01_Udemy_Course\05_Machine_Learning\08_Random_Forest\Data\clean\Travel_clean.csv", index=False)
