==> Nootbook/03_Feature_Engineering.ipynb <==
# Feature Engineering 


import pandas as pd 
import numpy as np
import matplotlib as plt
import seaborn as sns

#Importing Data
df = pd.read_csv(r"D:\Amaan_Behlim\01_Udemy_Course\05_Machine_Learning\09_ADABoost\Data\clean\Travel_clean.csv")
df.head()

### Feature Extraction

# Droping CustomerID
df = df.drop('CustomerID',axis=1)

#Creating new coloum for feature
df['TotalVisiting'] = df['NumberOfPersonVisiting'] + df['NumberOfChildrenVisiting']
df.drop(['NumberOfChildrenVisiting','NumberOfPersonVisiting'], axis=1, inplace=True)

df.info()

df.to_csv(r"D:\Amaan_Behlim\01_Udemy_Course\05_Machine_Learning\08_Random_Forest\Data\processed\Travel_processed.csv",index=False)

#Getting all categorical Features
categorical = [feature for feature in df.columns if df[feature].dtype =='str']
print(f'Categorical features: {len(categorical)}')
print(f'Names:\n{categorical}')
print('\n')

#Getting all numerical Features
numerical = [feature for feature in df.columns if df[feature].dtype !='str']
print(f'Numerical features: {len(numerical)}')
print(f'Names:\n{numerical}')
print("\n")

#Getting Discrete Features
descrete = [feature for feature in numerical if len(df[feature].unique())<=25]
print(f'Descrete features: {len(descrete)}')
print(f'Names:\n{descrete}')
print("\n")

#Getting Discrete Features
continous = [feature for feature in numerical if feature not in descrete]
print(f'Continous features: {len(continous)}')
print(f'Names:\n{continous}')
