==> Nootbook/04_Training.ipynb <==
# Model Training 
1. Spliting Dataset
2. Transforming Dataset
3. Model Training

### 1. Spliting Dataset 

import pandas as pd 
df = pd.read_csv(r"D:\Amaan_Behlim\01_Udemy_Course\05_Machine_Learning\09_ADABoost\Data\processed\Travel_processed.csv")
df.head()

from sklearn.model_selection import train_test_split

X = df.drop('ProdTaken',axis=1)
y = df['ProdTaken']

X_train, X_test, y_train, y_test = train_test_split(X,y, test_size=0.20, random_state=30)
X_train.shape, X_test.shape

### 2. Transforming Dataset

#Creatng Column Transformer with 3 types of transformer
catagorical_feature = X.select_dtypes(include='str').columns
numerical_feature = X.select_dtypes(exclude='str').columns

from sklearn.preprocessing import OneHotEncoder, StandardScaler
from sklearn.compose import ColumnTransformer

num_transformer = StandardScaler()
OH_transformer = OneHotEncoder(drop='first')

processor = ColumnTransformer(
    [
        ('OneHotEncoder', OH_transformer, catagorical_feature),
        ('StandardScalar', num_transformer, numerical_feature)
    ]
)

#applying Transformation on Training Dataset (fit_transform())
X_train = processor.fit_transform(X_train)
print("X Train")
print(X_train)
print("\n")

#applying Transformation on Test Dataset (transform())
X_test = processor.transform(X_test)
print("X Test")
print(X_test)
print("\n")

### 3. Training Dataset

import joblib
from sklearn.ensemble import RandomForestClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import AdaBoostClassifier
from sklearn.metrics import accuracy_score, f1_score, precision_score, recall_score, roc_auc_score

models = {
    "RandomForest":RandomForestClassifier(random_state=42),
    "Decisision_Tree":DecisionTreeClassifier(random_state=42),
    "Logistic Regression": LogisticRegression(random_state=42),
    "ADA Boost": AdaBoostClassifier(random_state=42)
}

for name, model in models.items():
    # model = list(model.values())[i]
    model.fit(X_train,y_train) #Train Model
    

    print(f"{name} model saved successfully!")


    # Make Prediction 
    y_train_pred = model.predict(X_train)
    y_test_pred = model.predict(X_test)

    # Training dataset Performance
    model_train_accuracy_score = accuracy_score(y_train,y_train_pred)
    model_train_f1_score = f1_score(y_train,y_train_pred) 
    model_train_precision_score = precision_score(y_train,y_train_pred)
    model_train_recall_score = recall_score(y_train,y_train_pred)
    model_train_roc_auc_score = roc_auc_score(y_train,y_train_pred)

    # Test Datase Performance
    model_test_accuracy_score = accuracy_score(y_test,y_test_pred)
    model_test_f1_score = f1_score(y_test,y_test_pred)
    model_test_precision_score = precision_score(y_test,y_test_pred)
    model_test_recall_score = recall_score(y_test,y_test_pred)
    model_test_roc_auc_score = roc_auc_score(y_test,y_test_pred)

    print(name)

    print("Model Performance For Training Set")

    print(f"Accuracy: {model_train_accuracy_score}")
    print(f"F1 Score: {model_train_f1_score}")
    print(f"Precision: {model_train_precision_score}")
    print(f"Recall: {model_train_recall_score}")
    print(f"ROC AUC: {model_train_roc_auc_score}")
    print("\n")


    print("Model Performance For Test Set")

    print(f"Accuracy: {model_test_accuracy_score}")
    print(f"F1 Score: {model_test_f1_score}")
    print(f"Precision: {model_test_precision_score}")
    print(f"Recall: {model_test_recall_score}")
    print(f"ROC AUC: {model_test_roc_auc_score}")
    print("\n")




# Hyperparameter Tunning
RandomForest_params = {
    "criterion": ['gini', 'entropy', 'log_loss'],
    "max_depth": [1,2,3,4,5,6,7,8,9,10,None],
    "min_samples_split": [2,18,15,20,],
    "max_features": ['sqrt','log2',None]
}
AdaBoost_params = {
    'n_estimators': [1,10,100],
    'learning_rate': [1.0,0.1,0.001,0.01],
    'n_estimators': [10,30,50,80]
}

randoncv_models = [
    ("Random Forest", RandomForestClassifier(),RandomForest_params),
    ("Ada Boost", AdaBoostClassifier(), AdaBoost_params)
]

from sklearn.model_selection import RandomizedSearchCV
model_params = {}

for name,model,params in randoncv_models:
    random = RandomizedSearchCV(estimator=model, 
                                cv=3, 
                                param_distributions=params, 
                                verbose=2, 
                                n_iter=100, 
                                n_jobs=-1)
    random.fit(X_train,y_train)
    model_params[name] = random.best_params_

    for model_name in model_params:
        print(f"------- Best Parameters for {model_name} ---------")
        print(model_params[model_name])

from sklearn.ensemble import RandomForestClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score, f1_score, precision_score, recall_score, roc_auc_score

models = {
    "RandomForest":RandomForestClassifier(min_samples_split = 2, max_features = None, max_depth = None, criterion = 'log_loss'),
    "Decisision_Tree":DecisionTreeClassifier(),
    "Logistic Regression": LogisticRegression(),
    "Ada Boost": AdaBoostClassifier(n_estimators=30, learning_rate=1.0)
}

for name, model in models.items():
    # model = list(model.values())[i]
    model.fit(X_train,y_train) #Train Model
    joblib.dump(model, rf"D:\Amaan_Behlim\01_Udemy_Course\05_Machine_Learning\08_Random_Forest\models/{name}.pkl") #saving model


    # Make Prediction 
    y_train_pred = model.predict(X_train)
    y_test_pred = model.predict(X_test)

    # Training dataset Performance
    model_train_accuracy_score = accuracy_score(y_train,y_train_pred)
    model_train_f1_score = f1_score(y_train,y_train_pred) 
    model_train_precision_score = precision_score(y_train,y_train_pred)
    model_train_recall_score = recall_score(y_train,y_train_pred)
    model_train_roc_auc_score = roc_auc_score(y_train,y_train_pred)

    # Test Datase Performance
    model_test_accuracy_score = accuracy_score(y_test,y_test_pred)
    model_test_f1_score = f1_score(y_test,y_test_pred)
    model_test_precision_score = precision_score(y_test,y_test_pred)
    model_test_recall_score = recall_score(y_test,y_test_pred)
    model_test_roc_auc_score = roc_auc_score(y_test,y_test_pred)

    print(name)

    print("Model Performance For Training Set")

    print(f"Accuracy: {model_train_accuracy_score}")
    print(f"F1 Score: {model_train_f1_score}")
    print(f"Precision: {model_train_precision_score}")
    print(f"Recall: {model_train_recall_score}")
    print(f"ROC AUC: {model_train_roc_auc_score}")
    print("\n")


    print("Model Performance For Test Set")

    print(f"Accuracy: {model_test_accuracy_score}")
    print(f"F1 Score: {model_test_f1_score}")
    print(f"Precision: {model_test_precision_score}")
    print(f"Recall: {model_test_recall_score}")
    print(f"ROC AUC: {model_test_roc_auc_score}")
    print("\n")
